refactor: Finalize Static File Organization
This commit marks the completion of the static file organization in the project. All static assets, including stylesheets, scripts, and images, have been reviewed, replaced, or newly added to ensure a polished and cohesive visual representation.

Key changes in this commit:

1. Stylesheets: Replaced and finalized CSS files in the 'css' directory.
2. Scripts: Updated and organized JavaScript files in the 'js' directory.
3. Images: Checked and placed all image assets within the 'images' directory.

This refactoring effort ensures consistency, improves maintainability, and enhances the overall structure of the project's static files.

+++ static/js/cart.js
@@ -7,14 +7,19 @@ for (var i = 0; i < updateBtns.length; i++) {
         var sizeId = this.dataset.sizeId
         var colorSizeId = this.dataset.color_size_id
         var action = this.dataset.action
-        var quantityInput = document.getElementById('qty');
+        var quantityInput = document.getElementById('qty-' + productId);
         var quantity = quantityInput ? quantityInput.value : 1;
 
+        if (action === 'delete_cart') {
+            if (!(confirm("آیا مطمئن هستید که می خواهید سبد خرید را پاک کنید؟"))) {
+                return;
+            }
+        }
         if (!(action === 'delete_item' || action === 'delete_cart')) {
             var intQtyInput = parseInt(quantity)
             if (intQtyInput > 0) {
                 action = 'add';
-                console.log('this is add')
+
             } else if (intQtyInput < 0) {
                 action = 'remove';
             } else {
@@ -26,12 +31,29 @@ for (var i = 0; i < updateBtns.length; i++) {
     })
 }
 
+function updateCart() {
+    var updateBtnsLazy = document.getElementsByClassName('update-cart-lazy')
+
+    for (var i = 0; i < updateBtnsLazy.length; i++) {
+        var button = updateBtnsLazy[i]
+        var productId = button.dataset.product
+        var colorSizeId = button.dataset.color_size_id
+        var action = button.dataset.action
+        var quantity = button.value;
+
+        if (parseInt(quantity) < 0) {
+            action = 'delete_item';
+        }
+
+        console.log('this is data: ', productId, colorSizeId, action, quantity,)
+        updateUserOrder(productId, null, null, colorSizeId, action, quantity)
+
+
+    }
+}
+
 function updateUserOrder(productId, colorId, sizeId, colorSizeId, action, quantity) {
-    console.log('user log in', colorSizeId)
-
-    var url = 'http://127.0.0.1:8000/products/cart/update-item/'
-
-    fetch(url, {
+    fetch(updateOrderUrl, {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
@@ -49,12 +71,10 @@ function updateUserOrder(productId, colorId, sizeId, colorSizeId, action, quanti
 
 
         .then((response) => {
-            console.log('data')
             return response.json()
         })
 
         .then((data) => {
-            console.log('data', data)
             location.reload()
         })
